# Vectorize ValidateHeaderCharacters