--- a/src/Microsoft.AspNetCore.Server.Kestrel/Internal/Http/FrameHeaders.Generated.cs
+++ b/src/Microsoft.AspNetCore.Server.Kestrel/Internal/Http/FrameHeaders.Generated.cs
@@ -7683,7 +7683,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Internal.Http
         }
         protected override void SetValueFast(string key, StringValues value)
         {
-            ValidateHeaderCharacters(value);
+            ValidateHeaderCharacters(ref value);
             switch (key.Length)
             {
                 case 13:
@@ -8025,7 +8025,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Internal.Http
         }
         protected override void AddValueFast(string key, StringValues value)
         {
-            ValidateHeaderCharacters(value);
+            ValidateHeaderCharacters(ref value);
             switch (key.Length)
             {
                 case 13: