x 1.9 times faster for plaintext (`13` + `text/plain`)
x 2.5 - 3.5 times faster for longer headers like `Content-Security-Policy` or large cookies

Newer version of part of #1000 "More Vector performance" for `ValidateHeaderCharacters`

```
        Method |                               Value |          Mean |     StdDev |            RPS |
-------------- |------------------------------------ |-------------- |----------- |--------------- |
  ValidateIter |                                  13 |     4.7229 ns |  0.0492 ns | 211,733,373.12 |
ValidateVector |                                  13 |     5.0420 ns |  0.0366 ns | 198,332,954.88 |
  ValidateIter |                          text/plain |    16.6734 ns |  0.3520 ns |  59,975,946.30 |
ValidateVector |                          text/plain |     6.2288 ns |  0.0638 ns | 160,543,964.49 |
  ValidateIter |             Content-Security-Policy |    37.6319 ns |  0.2950 ns |  26,573,199.60 |
ValidateVector |             Content-Security-Policy |    10.7913 ns |  0.1514 ns |  92,667,513.70 |
  ValidateIter | Mozilla/5.0 (Windows... 109 byte UA |   115.1069 ns |  3.0876 ns |   8,687,577.53 |
ValidateVector | Mozilla/5.0 (Windows... 109 byte UA |    36.4889 ns |  0.3958 ns |  27,405,596.46 |
  ValidateIter | ZGVmYXVsdC1zcmM... 1168 byte cookie | 1,047.3965 ns | 10.6608 ns |     954,748.25 |
ValidateVector | ZGVmYXVsdC1zcmM... 1168 byte cookie |   362.3385 ns |  2.9977 ns |   2,759,850.40 |
```

Testing at https://github.com/benaadams/PerfTesting/tree/master/src/HeaderValidation

/cc @halter73