@benaadams 2017-01-10
Updated with benchmarks (Vector-size 16, will be different for vector 32/64)