@benaadams 2017-01-23
Better testing using benchmark from https://github.com/aspnet/KestrelHttpServer/pull/1308

Before
```
        Method |                 Type |          Mean |        Median |           RPS |
-------------- |--------------------- |-------------- |-------------- |-------------- |
    SetHeaders |               Common | 1,435.1032 ns | 1,440.4368 ns |    696,814.01 |
    SetHeaders | ContentLengthNumeric |    84.2272 ns |    84.1386 ns | 11,872,651.77 |
    SetHeaders |  ContentLengthString |    58.5223 ns |    58.9710 ns | 17,087,502.97 |
    SetHeaders |            Plaintext |   154.9097 ns |   154.8979 ns |  6,455,373.13 |
    SetHeaders |              Unknown | 2,421.2750 ns | 2,427.1806 ns |    413,005.54 |
```
After
``` 
        Method |                 Type |          Mean |        Median |           RPS |
-------------- |--------------------- |-------------- |-------------- |-------------- |
    SetHeaders |               Common |   952.3173 ns |   955.1426 ns |  1,050,070.16 |
    SetHeaders | ContentLengthNumeric |    83.8107 ns |    83.8624 ns | 11,931,648.89 |
    SetHeaders |  ContentLengthString |    58.7320 ns |    58.7873 ns | 17,026,490.28 |
    SetHeaders |            Plaintext |   146.2217 ns |   146.1196 ns |  6,838,927.89 |
    SetHeaders |              Unknown | 1,694.4606 ns | 1,764.4501 ns |    590,158.32 |
```